*: update Cortex (and Prometheus) dependency
Update Cortex dependency (and Prometheus together) to include
https://github.com/cortexproject/cortex/commit/70dddb6b70b87f899ab115e79426a3ba522ae6ad.

storage.Querier finally has support for label.Matchers so now we can
fully pass them down.

+++ cmd/thanos/receive.go
@@ -695,7 +695,7 @@ type receiveConfig struct {
 	tsdbMinBlockDuration       *model.Duration
 	tsdbMaxBlockDuration       *model.Duration
 	tsdbAllowOverlappingBlocks bool
-	tsdbMaxExemplars           int
+	tsdbMaxExemplars           int64
 
 	walCompression bool
 	noLockFile     bool
@@ -773,7 +773,7 @@ func (rc *receiveConfig) registerFlag(cmd extkingpin.FlagClause) {
 		"Enables support for ingesting exemplars and sets the maximum number of exemplars that will be stored per tenant."+
 			" In case the exemplar storage becomes full (number of stored exemplars becomes equal to max-exemplars),"+
 			" ingesting a new exemplar will evict the oldest exemplar from storage. 0 (or less) value of this flag disables exemplars storage.").
-		Default("0").IntVar(&rc.tsdbMaxExemplars)
+		Default("0").Int64Var(&rc.tsdbMaxExemplars)
 
 	cmd.Flag("hash-func", "Specify which hash function to use when calculating the hashes of produced files. If no function has been specified, it does not happen. This permits avoiding downloading some files twice albeit at some performance cost. Possible values are: \"\", \"SHA256\".").
 		Default("").EnumVar(&rc.hashFunc, "SHA256", "")
